requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
ultralytics==8.3.88
mediapipe==0.10.14
transformers==4.44.2
sentence-transformers==3.2.1
easyocr==1.7.2
openai==1.50.2
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1